refactor: move getFeedstocks import to utils and add getDatasetRepr utility function
